Create messages module that looks after win messages
