+++ lib/messages.rb
@@ -1,0 +1,14 @@
+module Messages
+
+  def player1_won
+    "#{@players.name1} won!"
+  end
+
+  def player2_won
+    "#{@players.name2} won!"
+  end
+
+  def draw
+    "It's a draw!"
+  end
+end
